# contracts/openapi/admin.yaml
openapi: 3.0.3
info:
  title: YoBFF 控制面 API
  version: 1.0.0
servers:
  - url: http://localhost:9090
security:
  - BearerAuth: []
paths:
...
  /api/v1/login:
    post:
      summary: 管理员登录
      description: 验证用户名密码并返回 API Token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: password
                captcha_id:
                  type: string
                  description: 需要验证码时填写，由 /api/v1/captcha 获取
                  example: id-1
                captcha_code:
                  type: string
                  description: 需要验证码时填写，用户输入的验证码文本
                  example: "1234"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
  /api/v1/login/require-captcha:
    get:
      summary: 查询登录是否需要验证码
      description: 根据当前客户端失败状态判断是否要求验证码
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  required:
                    type: boolean
                    example: false
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
  /api/v1/logout:
    post:
      summary: 管理员登出
      description: 提示客户端清理令牌
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: logged_out
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
  /api/v1/captcha:
    get:
      summary: 获取验证码
      description: 生成一次性验证码图片（Base64），用于登录人机校验
      security: []
      parameters:
        - in: query
          name: preset
          schema:
            type: string
            enum:
              - easy
              - default
          required: false
          description: 验证码预设，easy 更易识读
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptchaResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "503":
          $ref: "#/components/responses/CaptchaUnavailable"
  /api/v1/config:
    get:
      summary: 获取当前配置
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
    put:
      summary: 更新并应用配置
      parameters:
        - in: header
          name: X-Operator
          schema:
            type: string
          required: false
          description: 操作人标识
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Config"
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: updated
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
  /api/v1/config/validate:
    post:
      summary: 预检配置合法性
      parameters:
        - in: header
          name: X-Operator
          schema:
            type: string
          required: false
          description: 操作人标识
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Config"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigValidationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
  /api/v1/config/versions:
    get:
      summary: 查询配置版本列表
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            format: int32
            default: 20
          required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigVersionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
  /api/v1/config/rollback:
    post:
      summary: 回滚到历史配置版本
      parameters:
        - in: header
          name: X-Operator
          schema:
            type: string
          required: false
          description: 操作人标识
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigRollbackRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigRollbackResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
  /api/v1/config/cdn:
    get:
      summary: 获取 CDN 同步配置与状态
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  config:
                    $ref: "#/components/schemas/CDNSyncConfig"
                  status:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/CDNStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
    put:
      summary: 更新 CDN 同步配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CDNSyncConfig"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: updated
                  config:
                    $ref: "#/components/schemas/CDNSyncConfig"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/config/reload:
    post:
      summary: 触发配置热重载
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: reloaded
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
...
  /api/v1/certs:
    get:
      summary: 获取 SSL 证书列表
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          required: false
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 15
          required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
    post:
      summary: 上传 SSL 证书
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                cert:
                  type: string
                  format: binary
                key:
                  type: string
                  format: binary
                name:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSLCertificate"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
...
  /api/v1/log/level:
    get:
      summary: 查询当前日志级别
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
          content:
            application/json:
              schema:
                type: object
                properties:
                  level:
                    type: string
                    example: info
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
    put:
      summary: 更新日志级别
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogLevelUpdate"
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: updated
                  level:
                    type: string
                    example: info
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
  /api/v1/log/stats:
    get:
      summary: 获取日志管线统计
      responses:
        "200":
          description: OK
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            X-RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/RateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/RateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  headers:
    RequestID:
      description: 请求追踪标识
      schema:
        type: string
    RateLimitLimit:
      description: 当前时间窗口最大请求数
      schema:
        type: integer
        format: int32
    RateLimitRemaining:
      description: 当前窗口剩余请求数
      schema:
        type: integer
        format: int32
    RateLimitReset:
      description: 窗口重置时间戳
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: 请求非法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 鉴权失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: 触发限流
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AuthUnavailable:
      description: 鉴权未配置
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    CaptchaUnavailable:
      description: 验证码服务不可用
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MethodNotAllowed:
      description: 请求方法不允许
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationFailed:
      description: 配置预检失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
...
    ConfigValidationResponse:
      type: object
      required:
        - valid
        - errors
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
...
    ConfigVersionResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ConfigVersion"
    ConfigRollbackRequest:
      type: object
      required:
        - version_id
      properties:
        version_id:
          type: string
...
    Config:
      type: object
      properties:
        dataPlane:
          $ref: "#/components/schemas/DataPlaneConfig"
        controlPlane:
          $ref: "#/components/schemas/ControlPlaneConfig"
        security:
          $ref: "#/components/schemas/SecurityConfig"
        routing:
          $ref: "#/components/schemas/RoutingConfig"
        cdnSync:
          $ref: "#/components/schemas/CDNSyncConfig"
        certificates:
          type: array
          items:
            $ref: "#/components/schemas/Certificate"
        sslCertificates:
          type: array
          items:
            $ref: "#/components/schemas/SSLCertificate"
...
    CaptchaResponse:
      type: object
      required:
        - captcha_id
        - image_base64
      properties:
        captcha_id:
          type: string
        image_base64:
          type: string
        preset:
          type: string
...